Filter archived datasets by default
Expand our plugin to implement IPackageController so we can make use of
the before_search method to inject an extra query parameter and ignore
datasets with the "archived" tag.

Add a button to the Search page filters block to hide/show archived
datasets based on a param in the URL.  Since params are converted to
search query items by CKAN we need to remove it from the generated query
and hide it from the active filters block.